Refactored
